refactor: move handler tests from key_bundle_watcher_test.go to singleton_test.go
key_bundle_watcher.go was deleted when the generic watcher was extracted
to pkg/filewatcher/. The remaining test file tests keyBundleHandler()
which is defined in singleton.go, so the tests belong in singleton_test.go.

Partially generated by AI.
